pyproject.toml
[tool.poetry]
name = "drf-iiif-store"
version = "0.1.0"
description = "Django REST Framework IIIF Store"
license = "MIT"
authors = ["Finlay McCourt <finlay.mccourt@digirati.com>"]
readme = "README.md"
packages = [
    { include = "iiif_store" },
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "4.0.2"
djangorestframework = "^3.13.1"
django-model-utils = "^4.2.0"
beautifulsoup4 = "^4.10.0"
bleach = "^4.1.0"
python-dateutil = "^2.8.2"
django-q = "^1.3.9"
drf-search-service = {git = "https://github.com/digirati-co-uk/drf-search-service", rev = "0.1"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
requests = "^2.27.1"
pytest-docker = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
